chore: Rename component
